requirements.txt
# BigQuery 2025 Clinical Trial Matching - Python Dependencies
# Competition: Kaggle BigQuery 2025 Hackathon
# Last Updated: September 2025

# Core Dependencies
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
google-cloud-storage==2.17.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1

# BigQuery 2025 Features
bigframes==1.13.0
db-dtypes==1.2.0
google-cloud-aiplatform==1.60.0
vertexai==1.60.0

# API Framework
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
httpx==0.27.0

# Data Processing
scikit-learn==1.5.0
scipy==1.13.1
pyarrow==16.1.0

# Visualization (for notebook)
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# Jupyter Support
jupyter==1.0.0
ipykernel==6.29.4
notebook==7.2.1
ipywidgets==8.1.3

# Utilities
tqdm==4.66.4
click==8.1.7
colorama==0.4.6
tabulate==0.9.0

# Testing (optional)
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# Development (optional)
black==24.4.2
pylint==3.2.3
mypy==1.10.0